Added prisma

+++ testcandidates/postgraphile/run_bench.sh
@@ -6,9 +6,12 @@ getMachineDNS () {
 
 cd provision
 terraform init
-terraform destroy --auto-approve
+if [[ $1 == "destroy" ]]
+then
+  terraform destroy --auto-approve
+fi
 terraform apply --auto-approve
 
-benchmarker=$(getMachineDns "postgraphile")
+benchmarker=$(getMachineDNS "postgraphile_benchmarker")
 
 ssh -i ~/.ssh/aws-bench.pem ubuntu@$benchmarker "cd ~ubuntu/aws-benchmarks/testcandidates/postgraphile && cat ../bench.yaml | docker run --name benchmarker -i --rm -p 8050:8050 -v $(pwd):/graphql-bench/ws hasura/graphql-bench:v0.3-warmup"
